+++ src/ua/data_types/relative_path.rs
@@ -3,8 +3,8 @@ use crate::ua;
 crate::data_type!(RelativePath);
 
 impl RelativePath {
-    #[allow(dead_code)] // This can be removed in the future
-    fn with_elements(mut self, elements: &[ua::RelativePathElement]) -> Self {
+    #[must_use]
+    pub fn with_elements(mut self, elements: &[ua::RelativePathElement]) -> Self {
         let array = ua::Array::from_slice(elements);
         array.move_into_raw(&mut self.0.elementsSize, &mut self.0.elements);
         self