+++ CHANGELOG.md
@@ -24,6 +24,8 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Add methods `Server::add_reference()` and `Server::delete_reference()`.
 - Add method `Error::status_code()` to get original OPC UA status code that caused the error.
 - Add method `ua::NodeId::into_expanded_node_id()`.
+- Add method `Server::translate_browse_path_to_node_ids()`.
+- Implement `Index` and `IndexMut` for `ua::Array` to allow direct element access.
 
 ### Changed